promql: fix unconvert issues (#4040)
See,
$ gometalinter --vendor --disable-all --enable=unconvert --deadline 6m ./...
promql/engine.go:1396:26:warning: unnecessary conversion (unconvert)
promql/engine.go:1396:40:warning: unnecessary conversion (unconvert)
promql/engine.go:1398:26:warning: unnecessary conversion (unconvert)
promql/engine.go:1398:40:warning: unnecessary conversion (unconvert)
promql/engine.go:1427:26:warning: unnecessary conversion (unconvert)
promql/engine.go:1427:40:warning: unnecessary conversion (unconvert)
promql/engine.go:1429:26:warning: unnecessary conversion (unconvert)
promql/engine.go:1429:40:warning: unnecessary conversion (unconvert)
promql/engine.go:1505:50:warning: unnecessary conversion (unconvert)
promql/engine.go:1573:46:warning: unnecessary conversion (unconvert)
promql/engine.go:1578:46:warning: unnecessary conversion (unconvert)
promql/engine.go:1591:80:warning: unnecessary conversion (unconvert)
promql/engine.go:1602:94:warning: unnecessary conversion (unconvert)
promql/engine.go:1630:18:warning: unnecessary conversion (unconvert)
promql/engine.go:1631:24:warning: unnecessary conversion (unconvert)
promql/engine.go:1634:18:warning: unnecessary conversion (unconvert)
promql/engine.go:1635:34:warning: unnecessary conversion (unconvert)
promql/functions.go:302:42:warning: unnecessary conversion (unconvert)
promql/functions.go:315:42:warning: unnecessary conversion (unconvert)
promql/functions.go:334:26:warning: unnecessary conversion (unconvert)
promql/functions.go:395:31:warning: unnecessary conversion (unconvert)
promql/functions.go:406:31:warning: unnecessary conversion (unconvert)
promql/functions.go:454:27:warning: unnecessary conversion (unconvert)
promql/functions.go:701:46:warning: unnecessary conversion (unconvert)
promql/functions.go:701:78:warning: unnecessary conversion (unconvert)
promql/functions.go:730:43:warning: unnecessary conversion (unconvert)
promql/functions.go:1220:23:warning: unnecessary conversion (unconvert)
promql/functions.go:1249:23:warning: unnecessary conversion (unconvert)
promql/quantile.go:107:54:warning: unnecessary conversion (unconvert)
promql/quantile.go:182:16:warning: unnecessary conversion (unconvert)
promql/quantile.go:182:64:warning: unnecessary conversion (unconvert)

+++ promql/quantile.go
@@ -104,7 +104,7 @@ func bucketQuantile(q float64, buckets buckets) float64 {
 		count -= buckets[b-1].count
 		rank -= buckets[b-1].count
 	}
-	return bucketStart + (bucketEnd-bucketStart)*float64(rank/count)
+	return bucketStart + (bucketEnd-bucketStart)*(rank/count)
 }
 
 // The assumption that bucket counts increase monotonically with increasing
@@ -179,5 +179,5 @@ func quantile(q float64, values vectorByValueHeap) float64 {
 	upperIndex := math.Min(n-1, lowerIndex+1)
 
 	weight := rank - math.Floor(rank)
-	return float64(values[int(lowerIndex)].V)*(1-weight) + float64(values[int(upperIndex)].V)*weight
+	return values[int(lowerIndex)].V*(1-weight) + values[int(upperIndex)].V*weight
 }
